refactor(folder structure) : move parser class
move the class with the parse role to the parser folder
