+++ src/parser/MimeParser.hpp
@@ -13,9 +13,9 @@
 #ifndef MIMEPARSER_HPP
 # define MIMEPARSER_HPP
 
-#include "Library.hpp"
+#include "../Library.hpp"
 #include "RequestParser.hpp"
-#include "File.hpp"
+#include "../File.hpp"
 
 class MimeParser
 {
